Use stream-lib 2.5.0 instead of internal fork.  (#84)
We forked stream-lib way back in the day.

There is no longer a reason to keep using the forked version as we no
longer use the forked classes, and the OSS upstream has long surpassed
our internal fork. This commit upgrades us to use stream-lib 2.5.0
instead of stream-lib 2.4.0-fork11252014. To ensure backwards
compatibility of HLL sketches generated by the forked version of the
repo, we add a test that attempts to read hardcoded sketches and asserts
that their cardinality is what we'd expect it to be.